refactored app home view
